AbstractMultiNodeTestSelfTest: Add workaround for broken database locking, fixes test to fail to prove current bug as it should
Before this change the following exception would only be printed on
stderr during test execution but would NOT result in test failure like
it should:
	Error while terminating plugin: java.lang.AssertionError
	java.lang.AssertionError
		at plugins.WebOfTrust.WebOfTrust.terminate(WebOfTrust.java:2151)
		at plugins.WebOfTrust.WebOfTrust.terminate(WebOfTrust.java:2133)
		at freenet.pluginmanager.PluginInfoWrapper.startShutdownPlugin(PluginInfoWrapper.java:146)
		at freenet.pluginmanager.PluginInfoWrapper.stopPlugin(PluginInfoWrapper.java:195)
		at freenet.pluginmanager.PluginManager.killPlugin(PluginManager.java:1147)
		at plugins.WebOfTrust.AbstractMultiNodeTestSelfTest.testTerminate(AbstractMultiNodeTestSelfTest.java:42)

The test will now fail as it should with:
	Testcase: testTerminate took 6.966 sec
		FAILED
	junit.framework.AssertionFailedError
		at plugins.WebOfTrust.AbstractMultiNodeTestSelfTest.testTerminate(AbstractMultiNodeTestSelfTest.java:46)

The failure is caused by the change to terminate() of commit
fe8d1555fa2664ee95afc7b4455813c63cf51044. This will be fixed in a
following commit.

Intentionally pushing this directly to branch next instead of putting it
onto a branch associated with a bugtracker issue:
The bugged commit wasn't released in a WoT build yet.
The primary motivation of a bugtracker entry would be to document the
issue for including it in the changelog of the next release, but since
the bug wasn't released yet that isn't necessary.