AbstractMultiNodeTest: Give more leeway with thread limit: 96 per node instead of 64; fix thread limit for 1 node
In at least one of my test runs of IntroductionClientTest I had observed
the previous thread limit to be exhausted so bumping this slightly is a
good idea.

The result is a limit of 480 threads for IntroductionClientTest which
is still less than the default of 500 which fred gives to a node on the
live network.
(We don't need more than the default because our nodes do not handle any
traffic except the usually handful of requests which the unit tests
create. A node on the live network probably has to handle hundreds of
requests concurrently.
Also all of the nodes share the thread pool.)