IdentityDownloader.storeTrustChangedCommand...(): Remove bogus FIXME aspect
The places where SubscriptionManager.storeTrustChangedCommand...() is
called *must* hold the said locks already:
- The lock for the table of Trust objects is class WebOfTrust, so the
lock on it will be held.
- Trust changes can induce Score changes and thus cause the "should we
fetch this Identity?" decisions to change, which would already in the
previous IdentityFetcher Identity downloading code require the lock on
the IdentityFetcher to be held in order to call its callbacks to tell
it about the changes. That lock is now the said
WebOfTrust.getIdentityDownloaderController() lock, so that locking
requirement is also met.
- Trust changes are a modification to the database so the transaction
lock must be held.