WebOfTrust.restoreOwnIdentity...(): Fix bogus workaround for issue #6596
The first for(Trust oldGivenTrust : oldGivenTrusts) loop was
accessing the ObjectSet after it was accessed already (by the ArrayList
copy constructor), which as explained by issue
https://bugs.freenetproject.org/view.php?id=6596
could result in not processing all the Trusts (albeit it may not have if
the constructor works different to a loop iterator).
This could cause the Trusts to not be deleted.

As the Trust.getTruster() Identity object is being deleted in the
related code that could cause Trust objects with a null pointer for the
truster to exist in old databases.
Thus we'll have to write database upgrade code to check for that as part
of deploying this patch. When doing so please check the build which
deployed the fix for issue #6596 for whether it already contains
suitable database upgrade code which can be re-used.
A bugtracker entry to write the upgrade code has been filed at:
https://bugs.freenetproject.org/view.php?id=7005