Travis CI: Try to fix the temporary deactivation of Travis CI on this branch
Fixup for 8ca3a75aebd11273b2f131c17a693586885ea7ac.

For some reason instead of disabling the build Travis would try to run
some ruby builder, which failed because we are no Ruby project and thus
provide no configuration for Ruby.
I suspect this is because the build-disabling code was happening right
at the beginning of the Travis script so the Travis environment was
lacking basic initialization and thus blindly assumed we're a Ruby
project.