IdentityDownloaderSlow: Complete & clarify FIXME
I've stared at the removed "- it leaves a time window..." part for over
an hour and still cannot fully comprehend what I had meant by that. I
was probably wrong with it:
Deduplication happens upon all files added to the queue mOutputQueue.
But downloads of the EditionHint queue in the database of the
IdentityDownloaderSlow are *only* deduplicated when deleteEditionHints()
is called for a edition higher than the duplicates. So by delaying that
call in the proposed approach we're actually maken the deduplication
*worse* as the duplicates will continue to exist in the queue for longer
and could thus lead to download attempts of them.
This is fixed by the proposal added by this commit of checking the
mOutputQueue for whether an edition was already donwloaded before
attempting to download it.