IdentityFetcher: Don't require usage of storeAbortFetchCommand...() / storeStartFetchCommand...() when restoring an Identity, only require the storeRestoreOwnIdentityCommand...()
I've changed my mind, it is a good idea to implement this here:
Changing all IdentityDownloader implementations to fulfill the job of
these callbacks in the storeRestoreOwnIdentity...() callback means that
we can change WebOfTrust.restoreOwnIdentity...() to only have to call
one callback upon the IdentityDownloader, not 3.

Also, as already apparent by this commit, dealing with all necessities
which arise due to restoring an OwnIdentity can be simplified if it is
implemented at each IdentityDownloader implementation:
The comment added by this commit shows that IdentityFetcher doesn't
have to call storeAbortFetchCommand...(), but that is an implementation
detail not apparent from the outside - so the benefit of not calling it
can only be reaped if we fully move all actions necessary to restore an
OwnIdentity to the storeRestoreOwnIdentityCommand...() callback.

See aaa6b35a8cdc83a5da76a77d49f0bdcf7dfc19fe the related FIXME.