IdentityDownloaderSlow.storeAbortFetchCommand...(): Review for safety for the purpose of Identity deletion; tweak things I spotted meanwhile
In preparation of using it in storeRestoreOwnIdentityCommand....() which
will happen for the purposes explained in the previous commit.

As said there the goal is to require outside classes to not have to call
storeAbortFetchCommand...() for restoring an OwnIdentity anymore so the
JavaDoc which is added  hereby is intentionally not added to the parent
interface specification of this function but to the implementation in
class IdentityDownloaderSlow as it will only be relevant to that class.