IdentityDownloaderFast.storeDeleteOwnIdentityCommand...(): Convert to new *Pre*/*Post* callbacks
The "if(shouldDownload(trustee)) storeAbortFetchCommand...()" has to be
removed because that asssumed that the function was being called *after*
the Score database was updated already - but the concept of this new
*Pre*RestoreOwnIdentity version of the function implies that the Score
database was not changed yet.

Thus we now infer whether to abort the downloads of the trustees from
the old state of the Score database, which is fortunately also possible.