IdentityFetcher.storeDeleteOwnIdentityCommand...(): Convert to new *Pre*/*Post* callbacks; implement
The concept of this is based on the implementations of those functions
at IdentityDownloaderFast.

Notice the important difference:
Where at IdentityDownloaderFast we iterate over the given Trusts of the
oldIdentity we here have to iterate over the given Scores.
This is because IdentityDownloaderFast only downloads Identitys which
have received a *direct* trust from an OwnIdentity, but IdentityFetcher
deals with downloading *all* Identitys which are trustworthy, including
those which have received *indirect* Trusts - which is represented by
the Score database.