Score: Documentation, in particular about IdentityDownloader's decisions
In preparation of shipping IdentityDownloader JavaDoc in the next
commit - that documentation can be more simple if the details of the
download decision are explained here.
And as Score is one of the very most important core classes this should
be documented here anyway.

Authorship documentation is removed as it hasn't been maintained in
years and can be obtained by the Git history anyway.