Travis CI: Use Gradle wrapper to fix failure on Oracle Java 9
Gradle would fail with:
    "Could not determine java version from '9.0.1'."

That's because the Gradle which Travis CI deploys is too old for Java 9:
https://github.com/travis-ci/travis-ci/issues/8975
(Stangely it did work with Java 9 some time ago when I filed the fred
PR we recently discussed, I don't know what they changed since.)

The Gradle wrapper will download the latest Gradle and thereby fix this.

In turn OpenJDK 7 is disabled because it doesn't work with the wrapper
anymore.
(I do still think it's not good that we don't test on Java 7 as that's
what we officially require for fred and what Ubuntu 14.04 deploys;
however testing on newer versions should be preferred to testing on
older versions.
That conflict didn't exist when I filed the aforementioned fred PR to
support Java 7. Now that Java 7 is in conflict with 9 I agree that it's
better to prefer testing with 9.)