AbstractJUnit3BaseTest.testClone(): Add debug code for investigating Java 9 failure
For https://bugs.freenetproject.org/view.php?id=7016

I need to run this on Travis CI because I cannot reproduce this locally
due to lack of Java 9 at hand so I'm pushing the debug code to the
repository.
It will anyway be helpful to have removal of this be part of the commit
which fixes the bug because adding the output to the commit message will
nicely visualize what's the issue here.