IdentityTest.testClone(): Fix failure on Java 9
Fixes:
> junit.framework.AssertionFailedError: static final boolean plugins.WebOfTrust.Persistent.$assertionsDisabled expected not same
>     at plugins.WebOfTrust.AbstractJUnit3BaseTest.testClone(AbstractJUnit3BaseTest.java:141)
>     at plugins.WebOfTrust.IdentityTest.testClone(IdentityTest.java:88)

Output of the hereby removed debug code on Java 7:
>  For debugging why assertNotSame() fails for $assertionsDisabled:
>  original is Boolean: true
>  clone is Boolean:    true
>  ==:       false
>  equals(): true
>  original hash:      2099400080
>  clone hash:         1460824350
>  Boolean.FALSE hash: 2041797946
>  Boolean.TRUE hash:  1856883009

On Java 9:
>  For debugging why assertNotSame() fails for $assertionsDisabled:
>  original is Boolean: true
>  clone is Boolean:    true
>  ==:       true
>  equals(): true
>  original hash:      224468764
>  clone hash:         224468764
>  Boolean.FALSE hash: 224468764
>  Boolean.TRUE hash:  1328349514

As you can see, the trigger of our bug is that Java 9's Field.get() was
changed to return the static constant Boolean.FALSE for the value
"false" of boolean fields instead of returning a fresh object of class
Boolean for every call.

The fix is to disable the containing if() for primitive types such as
boolean, int, etc. It didn't make sense to execute it for those anyway
as the point of the if() is to check whether member objects are
wrongly re-used by clone() instead of being cloned as well
- but primitive types are not objects.

Additionally code is added to ignore static fields as well as those
are the definition of intentional re-use.