Add db4o-7.4 as git submodule
In prepration of its upcoming removal from fred's freenet-ext.jar

This was done using:
  git submodule add https://github.com/xor-freenet/db4o-7.4.git
  git commit --all --gpg-sign

Notice that this hardcodes the used revision of that repository to
commit 371ecaea62f9d2ba1b447f93eef8641286aa60a4 (= tag v31):
If you do "git show" on this commit you can see that the commit hash of
db4o is hardcoded.
Changing it would require updating the submodule *and* committing that
update of the submodule revision to the WoT repository.

Thus anyone who clones WoT and downloads the submodule will get that
commit, not a later one. This means that I cannot maliciously alter db4o
in the future :) - which in turn allows Freenet to not have to host a
copy of the db4o repository.

Source: https://git-scm.com/book/en/v2/Git-Tools-Submodules

To obtain db4o after cloning WoT use:
  git submodule update --init
NOTICE: The following commit will amend the Ant builder to do this
automatically.