Ant builder: Fix JAR discovery for directory fred/build/output/
I hadn't noticed this failing by the Travis CI build for commit
e081cb54f3297edc09a6dddc9cb41be02c8882e4 because I forgot to delete the
Travis CI build cache, which had resulted in Travis still caching the
old locations of the JARs.
Sorry.

The <fileset> approach hereby chosen will NOT list anything in the path
if the JARs don't exist which is why the "it" is removed in the
print-libs task.