Travis CI: Re-enable & fix compiling on OpenJDK 7 so we can use the uploaded JAR for releases
Fred & WoT's minimal required Java version is Java 7.
Thus the most clean way to compile a release is on Java 7.
As this branch introduces automated upload of the resulting JAR to
Freenet it's hence a good idea to allow usage of that JAR for deployment
by hereby fixing Java 7 on Travis.