IdentityDownloader.storePostRestoreOwnIdentityCommand(): Documentation
As inspired by storePostDeleteOwnIdentiyCommand().

I'm adding this even though the hereby also amended FIXME looks like a
reason against adding it because IdentityDownloaderSlow's implementation
of the callback uses the Score database.
Guaranteeing the Score database to be up to date also means the Trust
database has to be update as Trusts are resulting from Scores, so adding
that as well.