IdentityDownloader: Resolve FIXME of wiring in storeTrustChangedCommand...()
I've reviewed all the said places where it needs to be wired in and it
is called in all of them already, no further changes required.

The places where it is not called were covered by the previous dozens of
commits which introduced store[Pre|Post]...() callbacks for handling
(Own)Identity deletion/restoring.

If I recall correctly this means the rough core of
IdentityDownloaderFast is done now, as well as IdentityDownloaderSlow
since some time ago, and what remains to be done is:
- review/resolve all FIXMEs added by this branch (important!!!)
- implementing start()/terminate() at IdentityDownloaderFast
- adding IdentityDownloaderFast to the StatisticsPage.