IdentityDownloaderFast: Implement terminate()
Based on IdentitFetcher.stop().
Not putting a hint at IdentityFetcher.stop()'s duplication of this in
the JavaDoc since that class will go away once the new
IdentityDownloader code is finished and tested.

An idea on how to avoid the huge code duplication among the different
stop() implementations has been added to the bugtracker:
https://bugs.freenetproject.org/view.php?id=6674#c12430