Add plotting library XChart as git submodule
In preparation of adding plotting of Identity file downloads vs time.

This was done using:
	git submodule add https://github.com/xor-freenet/XChart
	( cd XChart ; git verify-tag "$(git describe)" )
	git commit --all --gpg-sign

Notice that this hardcodes the used revision of that repository to
the head commit.
If you do "git show" on this commit you can see that the commit hash of
it is hardcoded.
Changing it would require updating the submodule *and* committing that
update of the submodule revision to the WoT repository.

Thus anyone who clones WoT and downloads the submodule will get that
commit, not a later one. This means that I cannot maliciously alter
XChart in the future :) - which in turn allows Freenet to not have to
host a copy of the XChart repository.

Source: https://git-scm.com/book/en/v2/Git-Tools-Submodules

To obtain XChart after cloning WoT use:
	git submodule update --init
NOTICE: The following commit will amend the Ant builder to do this
automatically.