Ant builder: Speed up build by over 1 minute to 12s; simplify
For some reason <archives> was very slow, thus replacing it with
<zipgroupfileset>.

Unforunately <zipgroupfileset> doesn't support nesting the <path>
element so we cannot use the existing <path id="submodules.path"> as
input for it.
As it supports using a directory as input with filename wildcards this
commit introduces a "dependencies/" directory to which it copies the
submodule JARs and from which it feeds the <zipgroupfileset>.

Fortunately this has the benefit of making introducing of further
dependencies even more easy because there are now less locations in the
Ant builder where each dependency needs to be listed.

The <path id="submodules.path"> is moved downwards so the related
comment of why <fileset> is used is above it.