Add class Pair
Apparently Java still doesn't provide this as of 2018.

Will be used by IdentityFileQueue.IdentityFileQueueStatistics to store
input for an X/Y plot.

Not obeying the WoT convention of prefixing the member variables with
"m" because this is subject to be replaced with an external library
class of e.g. Apache Java Commons once we bundle that.