XYChartUtils: Add movingAverage()
Roughly based on the above average().

The inversion of the if() condition was done since we only need the
else{} part here, the main part would be empty.
Notice that the inversion of the boolean which checks the seconds *is*
correct:
In average() we add the element which the current iteration discovered
to the average *inside* of the if(), but in movingAverage() we have
already added it *before* the if().
So where average()'s else-branch could only yield the element in the
next iteration after the current iteration's if() added the last missing
element to the average by executing the non-else branch, movingAverage()
can yield it in the current iteration of the if() already.
Also the boolean for seconds and for amount are treated differently
w.r.t. their inversion for a valid reason:
In average() the current X value which we compare to the seconds limit
has been increased before the if() already, but amount is increased
inside it.
In movingAverage() both are increased before the if().

The readability of average() may be improved by the next commit to avoid
having to understand the above text.