Class WebOfTrust: Remove leftover JavaDoc from reverted commit
The variable which this was supposed to annotate was added and removed
in 2010 by:
5aef78c4a9df262acf9451a87350f1e141194de9
8baf46789821ce8fdaeeaa8a8dc0b9390e2607fa

The previous commit has introduced code with the same goal, i.e.
detecting whether we're running in a unit test.
Nevertheless it implemented this different from the aforementioned
legacy commits:
It is using a system property which has to be passed to the JVM to
detect unit tests instead of changing a static variable in the
initialization of the unit tests.
This ensures the static variable can be final, and thus the compiler can
optimize production code to not constantly re-read the variable.
Also this allows usage of the final variable in defining constants in
other places of the code which aren't supposed to be re-read either
- that is what I'm planning to use it for anyway.