Gradle: Fix 'Could not find any version that matches com.android...'
Apparently it will try to download Gradle even if the local version is
sufficient, and as we don't specify a remote repository to download
stuff from this fails.

I don't care specifying a Gradle version to be compatible with anyway,
the build script should just be conservative enough so it works with any
Gradle available in a sufficiently recent distribution.