Gradle: Fix to not miss Version.class with "gradle clean jar"
Would previously not compile Version.class with "gradle clean jar".

The fix is to copy the Version.java before compiling it in a separate
task to the compile task.
If it was copied in the compileVersionTask, "gradle clean jar" would
wrongly assume the compileVersionTask is up to date and not run it.

The underlying reason may be that Gradle will assume a task to be up to
date if it has no input - and the input of the compileVersionTask would
be produced within it if this copying happened there, so the input
wouldn't exist before the task runs.