Gradle: Fix to include same files as Ant
The usecase of someone unpacking the JAR to view the GPL is
100% academic IMHO.
And the GPL is visible/bundled in all places where we publish JARs, no
need to put it here.

There's a technical reason for not bundling it though:
The Ant builder doesn't seem to bundle it either, so not bundling it
will simplify my upcoming script to verify the Gradle output against
Ant.