Gradle: Fix compileVersionFileTask to work incrementally
In order to ensure Version.class is rebuild if "git describe" output
changed its enough to have the input of the task change if the git
status changed.
That is the case because task prepareVersionFile is configured to always
execute, i.e. never be incremental.

Now all tasks should consider themselves as UP-TO-DATE if nothing
changed except tasks:
- prepareVersionFile
- jar

So the only task which wrongly isn't incremental is the jar task.
I cannot figure out why it isn't incremental from the "--debug" output,
so for now I will leave it as is. Perhaps it is just a bug of the Gradle
version I'm using.