XYChartUtils.movingAverage(): Don't check unrelated stuff in assert()
The typical input data for this function is measurements which happen at
random points in time, specifically whenever an Identity file download
succeeds we add a measurement, and the time when a download succeeds is
random.
So there is no correlation between the time and the amount of
measurements. Hence it makes no sense to compare them in an assert.

Besides we can replace this assert() with a proper upper boundary for
the average X value: As X is measuring the time, i.e. is linearly going
upward, the average must not be above the X value at the end of the
window.
Similarly, an assert for the lower boundary is also added.