XYChartUtils.differentiate(): Add TODO about strict mathematical correctness
Not having this be a FIXME because in our current usecase of this
function the underlying measurements will have deltas of the order of
magnitude of seconds to sub-seconds, but the rendering will be in the
order of minutes to hours.
I.e. even if this were incorrect as-is the resulting imprecision would
not be visible to the bare eye in the plot.