XYChartUtils.differentiate(): Improve accuracy of X-values of differential
In order to attempt to improve the position of the measured differential
on the resulting time charts, specifically the DownloadsPerHour plot of
StatisticsPage. This attempts to solve the issue of the end of the plot
seemingly lagging behind the end of the underlying TotalDownloadCount
plot. This can be spotted by an area of constant downloads per hour at
the end of the plot. It is constant due to the fact that the
StatiticsPage amends the data resulting from differentiation by adding
one last point as:
    downloadsPerHour.addLast(new Pair<>(currentTime,
        downloadsPerHour.peekLast().y));

I have not tested whether this improves upon the issue, however I would
anyway like to have the code changed like this because it makes more
sense:
Given that this differentiate() typically acts upon measurements, not
calculatable data, we do not really know whether the slope we measured
happened at the beginning or end of the measurement, so the best to
assume, given equal probability of all points in time, is that it
happened somewhere in the middle, which this commit changes the code
to.

The correctness of this can be deducted from:
https://en.wikipedia.org/w/index.php?title=Numerical_differentiation&oldid=854790732#Finite_difference_formulas

Specially, consider the statement:
> Another two-point formula is to compute the slope of a nearby secant
> line through the points (x-h,f(x-h)) and (x+h,f(x+h)). The slope of
> this line is (f(x+h) - f(x-h)) / 2h

In other words, the slope as we compute it from two measurements can be
considered as the slope of the function at the x-value which is in the
middle between the two measurements - which is what this commit change
the x-value to.