XYChartUtilsTest.testMovingAverage(): Add failing assert() to prove that the current test code is bogus
E.g. adding only 2 elements with 1 second spacing to the input data for
movingAverage() would give them a total spacing of 1 second.
Same applies to the current 32 elements, it gives a total spacing of 31
seconds, not 32.

Will fix this in the next commit, spent enough time today upon
discovering this.