XYChartUtils.movingAverage(): Fix to not unnecessarily yield a tail element
Just because the window contains >= 16 elements does not mean the main
loop did not yield an element for the last element(s) of it:
The window can grow beyond 16 elements if the 16 elements do not satisfy
the given amount of seconds which are required to yield an element.

While the goal during writing this was to fix the unit test failure it
does intentionally not do so:
The unit test is likely wrong, it would only succeed like this:
"if(unyieldedAmount >= 16 || result.size() == 0) {"

I.e. it tests for whether the window jumps by 16 elements for the tail
element - but we're a *moving* average, not a regular one. So the window
should move gradually by increments of one, not jump by 16.