Travis WoT JAR uploader: Fix shutdown to not cause errexit
Would previously fail with:
> cannot remove
> ‘/home/travis/build/xor-freenet/plugin-WebOfTrust/../fred/Freenet.anchor’:
> No such file or directory

This was possibly caused by not using the wrapper to start Freenet.

Not using the "--no-run-if-empty" as was previously used by the old
version of the script because the PID file should not be empty, if it is
then that is an error and we should notice it by kill complaining.