Travis CI: Run Ant on openjdk7, not 8, to make UI output less confusing
Now that we use the envvars FRED_BRANCH, BUILDER and DEPLOY they are each
displayed in the Travis UI in the job list of each build, next to the
Java version.
Since there are two builds on openjdk7, once for fred branch master and
once for next, there would be a job using Gradle between the first job
which is used for deployment, and the third which uses Ant.
So all the other jobs after Ant would look like the Gradle job in between,
which would make it seem like there is a reason Ant runs on 8, not 7.
But there isn't such a reason anymore, the only reason it was ran on 8 was
to ensure that deployment doesn't happen there back when there was no
DEPLOY envvar, which we have now to control where deployment happens.