XYChartUtils.movingAverage(): Remove fragile logic of yielding trailing elements
This isn't worth the hassle as the input dataset in typical usage of this
function of WoT (measuring downloads per hour) typically is much larger
than 16 elements, so the additional 1 output value won't be visible in the
resulting chart anyway.

This should fix the unit test for this function to succeed.