XYChartUtils: Remove movingAverageOld() in favor of the rewrite
The new code is probably fast enough:
For real world data sets it will probably run in O(const * N), where
const is the typical amount of measurements in the given constant amount
of seconds.
This is because for each of the N input data elements the inner loop only
runs backwards as far as necessary to yield enough measurements to cover
the amount of seconds.
The amount of measurements in those seconds will be a finite constant
because we usually use this function to analyze download statistics,
and download bandwidth is finite.