MathUtilTest: Fix bug which the previous commit introduced
The below for() loops need to skip the case where both i and j point to the
entry of 1d, but they check for index 0, not 6 which it was moved to by
the previous commit.
Fixing the position of 1d, not the if() in the loops, because "==0" is
easier to read than "==6".