IdentityDownloaderSlow: Fix AssertionError
queueSortedByDb4o is of type InitializingObjectSet, which has an assert to
detect and fail if it is being iterated over more than once.
Iterating over ObjectSets multiple times is currently not allowed due to
the db4o bug which is explained in the linked issue.

The specific AssertionError would be:

Could not load plugin [...]/WebOfTrust.jar : java.lang.AssertionError
java.lang.AssertionError
	at [...](Persistent.java:706)
	at [...](IdentityDownloaderSlow.java:1515)