IdentityDownloaderSlow: Fix wrong documentation, improve its readability
"positiveScore" is ambiguous because some people may consider zero as
positive and some may not.
That this was necessary is immediately apparent from the fix for the
wrong aspect of this comment:
An identity with a score of == 0 *is* considered as trustworthy, not
as a spammer - but the variable positiveScore does not include 0 as
positive.