XMLTransformer: Fix wrong assert and documentation
If you read the whole comment block before you'll see that
bestScore >= 0 is the case here, not > 0.
It wound up wrong because the variable "positiveScore" was previously used
here instead of "bestScore > 0" which is the content of that variable.
When I used positiveScore instead I had wrongly assumed that the == 0 case
is included in it being true.
The previous commits changed it to "bestScore > 0" to point out the
contrast in preparation of hereby fixing it.