IdentityDownloaderSlow: Move code out of deleteEditionHints() in preparation of calling it in onNewEditionImported()
The upcoming onNewEditionImported() implementation is called as part of
a database transaction which is managed by external code and thus it may
not commit the transaction - which deleteEditionHints() currently does,
so it will have to be changed to not commit.
It cites the hereby moved code as related to the committing.
Since the cited relationship between the commit and the code cannot be
obeyed anymore and thus isn't justification for the code to be there,
it is better to move it to onFailure() when considering it only on its
own:
Incrementing the failure counter is much more closely related to the
event of the download failing than to deletion of edition hints.

Note that this does change the behavior a bit:
Failures will now also be counted if the affected identity doesn't exist
anymore.
This is good though: Failures represent network traffic, and the more
network traffic we count the better the ability to observe potential
denial of service issues on the StatisticsPage of the web interface.