IdentityDownloaderSlow: Move code out of deleteEditionHints() in preparation of calling it in onNewEditionImported()
The upcoming onNewEditionImported() implementation is called as part of
a database transaction which is managed by external code and thus it may
not commit the transaction - which deleteEditionHints() currently does,
so it will have to be changed to not commit.
It cites the hereby moved code as related to the committing.
Since the cited relationship between the commit and the code cannot be
obeyed anymore and thus isn't justification for the code to be there,
it is better to move it to onSuccess() when considering it only on its
own:
Incrementing the skip-counter is much more closely related to the
event of the download succeeding than to deletion of edition hints,
especially because we only increment if downloadSucceeded == true.