XMLTransformer: Add member class ImportIdentityStatistics to measure execution time properly
In preparation of fixing the existing code at IdentityFileProcessor which
fails to exclude the time to acquire the locks.

Fixing this now because the recent commit
e8b0df96c6fd8cdda82651970e19ad6e3f0eea7e added a FIXME which requests to
do stuff based on the measurements.
Still I will probably not resolve the FIXME in the following commits
because  I want to get the release done - but I had written this code here
already so I'm gonna commit it on this branch anyway.