IdentityDownloaderSlow: Finish changes requested by large onSuccess() FIXME
The lock mLock isn't taken anymore in the place this commit moves code to
because the callers of that function will take the lock already as
required by its contract in interface IdentityDownloader.

A large amount of commits before this one did the many necessary changes
to allow this final one of commenting out the code in onSuccess():
- onNewEditionImported() was added to interface IdentityDownloader.
- Implementations of it have been added to all IdentityDownloaders.
  Notably the one of IdentityDownloaderSlow now contains the code which
  this commit comments out.
- XMLTransformer.importIdentity() was changed to call
  onNewEditionImported() for both succeeded and failed XML parsing.
- To ensure downloaded editions are not downloaded again in between
  onSuccess() and onNewEditionImported() the download scheduler of
  IdentityDownloaderSlow was changed to use
  IdentityFileQueue.containsAnyEditionOf() to ignore editions which are
  pending import. That function was also added by the large set of commits
  leading up to this one.

The FIXME isn't removed by this commit as it shall instead be recycled
into documentation. Given that it is a large piece of text this will be
dealt with by the following commits in small parts.