IdentityDownloaderSlow: Amend XMLTransformer commit ca97adfda59b9f46258337097f5f4d11bfc62a72
The said commit changed XMLTransformer to not import obsolete editions
anymore.

In turn, since download of obsolete editions is not normal expected
behavior anymore (and arguably never was, the code in XMLTransformer was
merely very lax, it wasn't that way because it would have been truly
expected), hereby an assert() is added to detect if the
IdentityDownloaderSlow wrongly would start a download of an obsolete
edition.