Revert "add translations, thanks to the team at transifex"
This reverts commit 0d15228ded52b735720f780d87fbf32050218b37.

Reverting because the following happened:

Arne had pushed the commit to branch next on Freenet's repo *without*
first submitting it to me for review.
But I cannot review it and thus cannot merge it:
Reviewing would include pulling the l10n files from Transifex myself to
see if what Arne committed matches what I get when I download them.
But I cannot download them, I lack access to the Transifex repo.
Arne has tried to get me access, but that has been taking a year now.
And I cannot wait any further for this:
I need to push other changes to branch next. And those need to be
fast-forwardable onto Freenet's branch next as per our policy of the
Freenet repo being a strict mirror of my repo.
But they won't be fast-forwardable while Freenet's next contains changes
which I haven't merged.

So the resolution is to merge a branch into my repo which contains the
affected commit of Arne, along with this revert commit which immediately
reverts it.
Then git correctly thinks that our repos are in sync, but the actual
changes of the l10n aren't applied.