GitHub Actions: Move disabled caching code to a better place
Caching consumes the result of compilation so it is better to put it
after compilation to define the result before it is consumed.

Further, it is the less interesting piece of code so it should be
further down in the file.

And finally, the following commits shall merge the code right below
where it was placed previously with the code right above where it was
placed, so this will improve the readability of their diffs.

Based on Freetalk commit cb3eea95f8930bc5e860b7df2a4e57134eea6f2b