Revert "GitHub Actions: Try to fix pyFreenet3 setup by installing python3-setuptools"
This reverts commit 2af536574c1920a075738bf6942e45d952f2249b.

While it did improve the issue, it did not fully fix it, now
installation would fail with a different error message.

Since debugging pyFreenet3 is out of my scope and this branch is large
enough already I'm postponing fixing this to a separate branch.

Reverting the installation of python3-setuptools because I have not yet
received a reply by Arne which would tell me if it is indeed something
which should be needed to be able to install pyFreenet3.
Will un-revert this if it is.