GitHub Actions: Amend the previous commit: Another fix attempt for Gradle ignoring the version we tell it to download
The ./gradle/ directory which we now delete is used to contain JARs of
the wrapper, so perhaps those are somehow stale copies of the 7.3.3
JARs.
Source:
https://docs.gradle.org/7.3.3/userguide/gradle_wrapper.html#sec:adding_wrapper

I don't know how they would get there given that we don't have caching
code in our GitHub Actions script yet, but perhaps GHA will just put
them there or whatever.

Even if this doesn't fix the issue it is still a bugfix we need:
The rm command otherwise would be wrong once we enable caching.