CurrentTimeUTC: Move the class from fred to WoT
The fred developers recently revamped the class, and then decided
to deprecate the class in fred, hence WoT (and  Freetalk) will have to
contain it so it can be removed there.

Its commit history was preserved by committing all of the history to the
WoT repository, which yielded the 4 commits before this one here.
This was done as such:

- On fred, use the following command to create *.patch files.
  Enter the command as a single line:

  $ git format-patch --output-directory ~/WebOfTrust/patches
  "e0d3a2ed19e282f960453f413d5f91a38d6cb161~1..
  286b5c98ff6d53b9c6dabd9753bfef13aaa8c781"
  src/freenet/support/CurrentTimeUTC.java

- On WoT, apply the patches with:

  $ git am --whitespace=nowarn patches/*

This added the class as of commit
  286b5c98ff6d53b9c6dabd9753bfef13aaa8c781
which was the last commit before the fred developers applied
modifications to it.

I am not including these modifications yet because I have not reviewed
them yet and have no time to do so right now.
I will document a TODO to do this either in the following commmits or in
the bugtracker.